Allow showing descriptions as tooltips
