+++ packages/ui/src/TapToEdit.tsx
@@ -3,11 +3,12 @@ import {Linking} from "react-native";
 
 import {Box} from "./Box";
 import {Button} from "./Button";
-import {BoxProps, FieldProps} from "./Common";
+import {TapToEditProps} from "./Common";
 import {Field} from "./Field";
 import {Icon} from "./Icon";
 import {useOpenAPISpec} from "./OpenAPIContext";
 import {Text} from "./Text";
+import {Tooltip} from "./Tooltip";
 
 export function formatAddress(address: any, asString = false): string {
   let city = "";
@@ -47,28 +48,6 @@ export function formatAddress(address: any, asString = false): string {
   }
 }
 
-export interface TapToEditProps extends Omit<FieldProps, "onChange" | "value"> {
-  title: string;
-  value: any;
-  // Not required if not editable.
-  setValue?: (value: any) => void;
-  // Not required if not editable.
-  onSave?: (value: any) => void | Promise<void>;
-  // Defaults to true
-  editable?: boolean;
-  // enable edit mode from outside the component
-  isEditing?: boolean;
-  // For changing how the non-editing row renders
-  rowBoxProps?: Partial<BoxProps>;
-  transform?: (value: any) => string;
-  fieldComponent?: (setValue: () => void) => ReactElement;
-  withConfirmation?: boolean;
-  confirmationText?: string;
-  confirmationHeading?: string;
-  openApiModel?: string;
-  openApiField?: string;
-}
-
 export const TapToEdit = ({
   value,
   setValue,
@@ -83,16 +62,18 @@ export const TapToEdit = ({
   withConfirmation = false,
   confirmationText = "Are you sure you want to save your changes?",
   confirmationHeading = "Confirm",
+  description: propsDescription,
   openApiModel,
   openApiField,
+  showDescriptionAsTooltip = false,
   ...fieldProps
 }: TapToEditProps): ReactElement => {
   const [editing, setEditing] = useState(false);
   const [initialValue] = useState(value);
   const {getModelField} = useOpenAPISpec();
 
-  let description: string | undefined;
-  if (openApiModel && openApiField) {
+  let description: string | undefined = propsDescription;
+  if (!description && openApiModel && openApiField) {
     description = getModelField(openApiModel, openApiField)?.description;
   }
 
@@ -204,6 +185,15 @@ export const TapToEdit = ({
     };
     const isClickable = fieldProps?.type === "url" || fieldProps?.type === "address";
 
+    const renderTitleDescription = (): React.ReactElement => {
+      return (
+        <Box flex="grow">
+          <Text weight="bold">{title}:</Text>
+          {Boolean(description && !showDescriptionAsTooltip) && <Text>{description}</Text>}
+        </Box>
+      );
+    };
+
     return (
       <Box
         direction="row"
@@ -213,10 +203,13 @@ export const TapToEdit = ({
         width="100%"
         {...rowBoxProps}
       >
-        <Box flex="grow">
-          <Text weight="bold">{title}:</Text>
-          {description && <Text>{description}</Text>}
-        </Box>
+        {showDescriptionAsTooltip ? (
+          <Tooltip idealDirection="top" text={description}>
+            {renderTitleDescription()}
+          </Tooltip>
+        ) : (
+          renderTitleDescription()
+        )}
         <Box direction="row" justifyContent="start" marginLeft={2}>
           <Box justifyContent="start" onClick={isClickable ? openLink : undefined}>
             <Text align="right" underline={isClickable}>
